pyproject.toml
[project]
name = "tobs"
version = "1.0"
description = "Telegram exporter to Markdown format"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiofiles==24.1.0",
  "aiohttp==3.12.15",
  "aiohttp-proxy==0.1.2",
  "aiorwlock==1.4.0",
  "aiosignal==1.4.0",
  "annotated-types==0.7.0",
  "attrs==25.3.0",
  "beautifulsoup4==4.13.5",
  "certifi==2025.8.3",
  "charset-normalizer==3.4.3",
  "colorama==0.4.6",
  "ffmpeg-python==0.2.0",
  "frozenlist==1.7.0",
  "idna==3.10",
  "loguru==0.7.3",
  "lz4>=4.3.3",
  "magic-filter==1.0.12",
  "markdown-it-py==4.0.0",
  "mdurl==0.1.2",
  "msgpack>=1.0.0",
  "multidict==6.6.4",
  "pillow==11.3.0",
  "propcache==0.3.2",
  "psutil==7.0.0",
  "pyaes==1.6.1",
  "pyasn1==0.6.1",
  "pydantic==2.11.7",
  "pydantic-core==2.33.2",
  "pygments==2.19.2",
  "pysocks==1.7.1",
  "python-dotenv==1.1.1",
  "redis==6.4.0",
  "requests==2.32.5",
  "rich==14.1.0",
  "rsa==4.9.1",
  "soupsieve==2.8",
  "telethon==1.41.0",
  "typing-extensions==4.15.0",
  "urllib3==2.5.0",
  "win32-setctime==1.2.0; sys_platform == 'win32'",
  "yarl==1.20.1",
  "faster-whisper>=1.1.0",
  "orjson>=3.11.4",
  "uvloop>=0.19.0; sys_platform != 'win32'",
  "cryptg>=0.4.0",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
mypy_path = "stubs"

# Ignore missing imports for external libraries without type stubs
[[tool.mypy.overrides]]
module = [
    "ffmpeg.*",
    "aiohttp_proxy.*",
    "aiorwlock.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests requiring external dependencies",
    "unit: marks tests as unit tests (isolated, fast)",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
